requirements.txt
pyyaml
semver
